+++ src/main/java/dev/dracarys/com/hospitalquerysystem/controller/StayController.java
@@ -1,9 +1,14 @@
 package dev.dracarys.com.hospitalquerysystem.controller;
 
+import dev.dracarys.com.hospitalquerysystem.dominio.Doctor;
+import dev.dracarys.com.hospitalquerysystem.dominio.Patients;
+import dev.dracarys.com.hospitalquerysystem.dominio.Stay;
 import dev.dracarys.com.hospitalquerysystem.requests.stay.StayDeleteRequestBody;
 import dev.dracarys.com.hospitalquerysystem.requests.stay.StayGetReturnObject;
 import dev.dracarys.com.hospitalquerysystem.requests.stay.StayPostRequestBody;
 import dev.dracarys.com.hospitalquerysystem.requests.stay.StayPutRequestBody;
+import dev.dracarys.com.hospitalquerysystem.service.DoctorServices;
+import dev.dracarys.com.hospitalquerysystem.service.PatientServices;
 import dev.dracarys.com.hospitalquerysystem.service.StayServices;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -17,6 +22,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/v1/hospital/stay")
@@ -26,18 +32,39 @@ public class StayController {
 
     private final StayServices stayServices;
 
+    private final DoctorServices doctorServices;
+
+    private final PatientServices patientServices;
+
     @PostMapping("/create")
     @Operation(summary = "Create new stay", description = "To perform the request, it is necessary to have the permission of (HEADNURSE / ATTENDANT)", tags = {"STAYS"})
     @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Success created new stay", content = { @Content(mediaType = "application/json", schema = @Schema(implementation = StayPostRequestBody.class))}),
-            @ApiResponse(responseCode = "409", description = "Already exists stay"),
-            @ApiResponse(responseCode = "404", description = "if doctor or patient not exist"),
-            @ApiResponse(responseCode = "400", description = "if body is incorrect"),
+            @ApiResponse(responseCode = "201", description = "Success created new stay", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = StayPostRequestBody.class))}),
+            @ApiResponse(responseCode = "400", description = "Already exists stay"),
+            @ApiResponse(responseCode = "400", description = "doctor or patient not exist"),
+            @ApiResponse(responseCode = "409", description = "body is incorrect"),
             @ApiResponse(responseCode = "401", description = "Unauthorized request")
     })
-    public ResponseEntity<Object> createNewStay(@RequestBody StayPostRequestBody stayPostRequestBody){
-        return stayServices.createNewStay(stayPostRequestBody);
+    public ResponseEntity<Object> createNewStay(@RequestBody StayPostRequestBody stayPostRequestBody) {
+
+        Optional<Doctor> doctorExist = doctorServices.findByCrm(stayPostRequestBody.getCrmDoctor());
+        Optional<Patients> patientExist = patientServices.findByCPF(stayPostRequestBody.getCpfPatient());
+
+        if (doctorExist.isPresent() && patientExist.isPresent()) {
+            Optional<Stay> stay = stayServices.findStayByDoctorAndPatient(doctorExist.get(), patientExist.get());
+
+            if (stay.isPresent()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Stay already exist");
+            }
+            stayServices.createNewStay(stayPostRequestBody);
+            return ResponseEntity.status(HttpStatus.CREATED).body("Stay created successfully");
+        }
+
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("Error check the parameters and try again later");
+
     }
+
 
     @GetMapping("/list")
     @Operation(summary = "List all stays", description = "To perform the request, it is necessary to have the permission of (HEADNURSE / ATTENDANT / DOCTOR)", tags = {"STAYS"})
@@ -45,29 +72,66 @@ public class StayController {
             @ApiResponse(responseCode = "200", description = "Success list all stays"),
             @ApiResponse(responseCode = "401", description = "Unauthorized request")
     })
-    public ResponseEntity<List<StayGetReturnObject>> findAll(){
+    public ResponseEntity<List<StayGetReturnObject>> findAll() {
         return new ResponseEntity<>(stayServices.listAllStays(), HttpStatus.OK);
     }
+
 
     @PutMapping("/edit")
     @Operation(summary = "Edit stay", description = "To perform the request, it is necessary to have the permission of (HEADNURSE / ATTENDANT)", tags = {"STAYS"})
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201", description = "Stay edited successfully"),
-            @ApiResponse(responseCode = "404", description = "Stay not found"),
-            @ApiResponse(responseCode = "401", description = "Unauthorized request")
+            @ApiResponse(responseCode = "400", description = "Stay not found"),
+            @ApiResponse(responseCode = "401", description = "Unauthorized request"),
+            @ApiResponse(responseCode = "409", description = "Body is incorrect")
     })
-    public ResponseEntity<Object> editStay(@RequestBody StayPutRequestBody stayPutRequestBody){
-        return stayServices.editStayInfo(stayPutRequestBody);
+    public ResponseEntity<Object> editStay(@RequestBody StayPutRequestBody stayPutRequestBody) {
+
+        Optional<Doctor> doctorExist = doctorServices.findByCrm(stayPutRequestBody.getCrmDoctor());
+
+        Optional<Patients> patientExist = patientServices.findByCPF(stayPutRequestBody.getCpfPatient());
+
+        if (doctorExist.isPresent() && patientExist.isPresent()) {
+            Optional<Stay> stayExist = stayServices.findStayByDoctorAndPatient(doctorExist.get(), patientExist.get());
+
+            if (stayExist.isPresent()) {
+
+                stayServices.editStayInfo(stayPutRequestBody);
+                return ResponseEntity.status(HttpStatus.OK).body("Stay details updated successfully");
+
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Stay wasn't found");
+        }
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("Check parameters and try again");
     }
+
 
     @DeleteMapping("/delete")
     @Operation(summary = "Delete stay", description = "To perform the request, it is necessary to have the permission of (HEADNURSE)", tags = {"STAYS"})
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Stay edited successfully"),
-            @ApiResponse(responseCode = "404", description = "Stay not found"),
+            @ApiResponse(responseCode = "400", description = "Stay not found"),
             @ApiResponse(responseCode = "409", description = "Body is incorrect")
     })
-    public ResponseEntity<Object> deleteStay(StayDeleteRequestBody stayDeleteRequestBody){
-        return stayServices.deleteStay(stayDeleteRequestBody);
+    public ResponseEntity<Object> deleteStay(StayDeleteRequestBody stayDeleteRequestBody) {
+
+        Optional<Doctor> doctorExist = doctorServices.findByCrm(stayDeleteRequestBody.getCrmDoctor());
+
+        Optional<Patients> patientExist = patientServices.findByCPF(stayDeleteRequestBody.getCpfPatient());
+
+        if (doctorExist.isPresent() && patientExist.isPresent()) {
+            Optional<Stay> stayExist = stayServices.findStayByDoctorAndPatient(doctorExist.get(), patientExist.get());
+
+            if (stayExist.isPresent()) {
+
+                stayServices.deleteStay(stayDeleteRequestBody);
+                return ResponseEntity.status(HttpStatus.OK).body("Stay delete successfully");
+
+            }
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Stay wasn't found");
+        }
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("Check parameters and try again");
     }
 }
